interop[py]: dynamically import available interoper modules
